backend/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
django-environ==0.10.0
dj-database-url==2.0.0
mysqlclient